Added grouping items by category and fixed data mismatch between Dtos and Models.

+++ Lombard.BL/Services/ITransactionsService.cs
@@ -8,8 +8,8 @@ namespace Lombard.BL.Services
 {
     public interface ITransactionsService
     {
-        Task<Transaction> BuyAsync(int itemId, int customerId, int quantity, decimal price);
-        Task<Transaction> SellAsync(int itemId, int customerId, int quantity, decimal price);
+        Task<Transaction> BuyAsync(int itemId, int customerId, float quantity, decimal price);
+        Task<Transaction> SellAsync(int itemId, int customerId, float quantity, decimal price);
         Task UpdateTransactionAsync(Transaction transaction);
         Task DeleteTransactionAsync(int id);
     }
